Allow processing commands before input file processing.
When setting the include dirs for the top-level read_verilog like
   -p 'verilog_defaults -add -I<path>'

the include dirs will be ignored if the script file is specified as a
normal input file, because passes_commands are executed after input file
processing.

Changing the order how -p commands and the input file are processed
seems to break existing stuff, hence this patch adds a new option -r for
this purpose.